Add unmodified previous configuration
This is an ugly mix of vimscript, lua, and lots of plugins that aren't
used alongside keybindings long forgotten, and not discoverable because
we're lacking a description and `which-keys`.

The goal is to incrementally remove those as we are aproaching a new
configuration, discarding things we believe we don't need and replace
others with the shiny new lua implementation.